deepface/api_v2/requirements.api.txt
SQLAlchemy
typing-extensions
fastapi
pydantic
starlette
email-validator
pydantic-settings
python-multipart
uvicorn[standard]
websockets
sqlmodel
pyjwt
cryptography
passlib
psycopg
psycopg2-binary
tenacity
requests>=2.27.1
numpy>=1.14.0
pandas>=0.23.4
gdown>=3.10.1
tqdm>=4.30.0
Pillow>=5.2.0
opencv-python>=4.5.5.64
tensorflow>=1.9.0
keras>=2.2.0
mtcnn>=0.1.0
retina-face>=0.0.1